+++ src/save.c
@@ -49,8 +49,6 @@ extern u32 sub_80954CC(void* a, s32 b);
 extern u32 sub_8095624(u8 *, u32);
 extern u32 RestoreMailInfo(void* a, s32 b);
 extern u32 SaveMailInfo(u8 *, u32);
-extern u32 sub_8097D60(u8 *, u32);
-extern u32 sub_8097D98(void* a, s32 b);
 extern void sub_800135C(void);
 extern void sub_80958E4(u32 *a, u32 b);
 extern void sub_80972F4(void);
